js: move Poller to ./Legacy.js
