+++ js/main.js
@@ -40,6 +40,7 @@ import ReactResizeDetector from 'react-resize-detector';
 
 import EmbeddingsPane from './EmbeddingsPane';
 import ImagePane from './ImagePane';
+import Poller from './Legacy';
 import NetworkPane from './NetworkPane';
 import PlotPane from './PlotPane';
 import PropertiesPane from './PropertiesPane';
@@ -104,101 +105,6 @@ if (ACTIVE_ENV !== '') {
 } else {
   use_env = localStorage.getItem('envID') || 'main';
   use_envs = JSON.parse(localStorage.getItem('envIDs')) || ['main'];
-}
-
-function postData(url = ``, data = {}) {
-  return fetch(url, {
-    method: 'POST',
-    mode: 'cors',
-    cache: 'no-cache',
-    credentials: 'same-origin',
-    headers: {
-      'Content-Type': 'application/json; charset=utf-8',
-    },
-    redirect: 'follow',
-    referrer: 'no-referrer',
-    body: JSON.stringify(data),
-  });
-}
-
-class Poller {
-  /**
-   * Wrapper around what would regularly be socket communications, but handled
-   * through a POST-based polling loop
-   */
-  constructor(app) {
-    this.app = app;
-    var url = window.location;
-    this.target =
-      url.protocol + '//' + url.host + app.correctPathname() + 'socket_wrap';
-    this.handleMessage = app._handleMessage;
-    fetch(this.target)
-      .then((res) => {
-        return res.json();
-      })
-      .then((data) => {
-        this.finishSetup(data.sid);
-      });
-  }
-
-  finishSetup = (sid) => {
-    this.sid = sid;
-    this.poller_id = window.setInterval(() => this.poll(), POLLING_INTERVAL);
-    this.app.setState({
-      connected: true,
-    });
-  };
-
-  close = () => {
-    this.app.setState({ connected: false }, () => {
-      this.app._socket = null;
-    });
-    window.clearInterval(this.poller_id);
-  };
-
-  send = (msg) => {
-    // Post a messge containing the desired command
-    postData(this.target, { message_type: 'send', sid: this.sid, message: msg })
-      .then((res) => res.json())
-      .then(
-        (result) => {
-          if (!result.success) {
-            this.close();
-          } else {
-            this.poll(); // Get a response right now if there is one
-          }
-        },
-        (error) => {
-          console.log(error);
-          this.close();
-        }
-      );
-  };
-
-  poll = () => {
-    // Post message to query possible socket messages
-    postData(this.target, { message_type: 'query', sid: this.sid })
-      .then((res) => res.json())
-      .then(
-        (result) => {
-          if (!result.success) {
-            this.close();
-          } else {
-            let messages = result.messages;
-            messages.forEach((msg) => {
-              // Must re-encode message as handle message expects json
-              // in this particular format from sockets
-              // TODO Could refactor message parsing out elsewhere.
-              this.handleMessage({ data: msg });
-            });
-          }
-        },
-        (error) => {
-          console.log(error);
-          this.close();
-        }
-      );
-  };
 }
 
 // TODO: Move some of this to smaller components and/or use something like redux
@@ -382,17 +288,13 @@ class App extends React.Component {
     }
   };
 
-  setupPolling = () => {
-    this._socket = new Poller(this);
-  };
-
   connect = () => {
     if (this._socket) {
       return;
     }
     // eslint-disable-next-line no-undef
     if (USE_POLLING) {
-      this.setupPolling();
+      this._socket = new Poller(this);
       return;
     }
     var url = window.location;
